+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: The parent folder will be created, if it does not exists
+
 ## 3.2.0 - 2019-08-29
 
 * Added: The Path parameter for Start-ScriptLogger can now be an empty string